moved Connections class to classes.py for consistency
